feat: add `Tooltip` component
