--- .github/ISSUE_TEMPLATE/feature.md
@@ -1,20 +1,0 @@
----
-name: Feature
-about: Feature Request
-title: "[F]"
-labels: feature
-assignees: ''
-
----
-
-## Expected Behavior
-
-
-## Actual Behavior
-
-
-## Steps to Reproduce the Problem
-
-  1.
-  1.
-  1.
