--- components/charts/Bar/styles.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Bar = styled.rect`
-  cursor: pointer;
-  fill: #30e0e3;
-  transform-origin: center bottom;
-`;
