--- components/charts/YAxis/styles.js
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-export const GridLine = styled.line`
-  stroke: var(--white, #fff);
-`;
-
-export const GridLineLabel = styled.text`
-  font-size: 10px;
-  alignment-baseline: mathematical;
-  dominant-baseline: mathematical;
-`;
