--- components/charts/index.js
@@ -1,2 +1,0 @@
-export { default as ChartBase } from "./ChartBase";
-export { default as ChartFigure } from "./ChartFigure";
