+++ components/molecules/InfoIcon/styles.ts
@@ -1,6 +1,12 @@
 import styled from "styled-components";
 
 export const InfoIcon = styled.div`
+  position: relative;
+  display: inline-block;
   grid-column: 2;
   place-self: center center;
+`;
+
+export const InfoIconContainer = styled.div`
+  position: relative;
 `;
