+++ components/svg/charts/ChartBase/styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const SVG = styled.svg`
+  width: 100%;
+  min-width: var(--min-width);
+  max-height: 100%;
+  aspect-ratio: var(--aspect-ratio);
+  margin-inline: auto;
+`;
