+++ components/svg/charts/LineData/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Line = styled.polyline`
+  fill: none;
+  stroke: #f9dc86;
+  stroke-width: 4;
+`;
