+++ components/svg/charts/XAxis/styles.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const BaseLine = styled.line`
+  stroke: var(--white, #fff);
+`;
+
+export const Tick = styled.line`
+  stroke: var(--white, #fff);
+  stroke-width: 2;
+`;
+
+export const TickLabel = styled.text`
+  font-size: 14px;
+  text-anchor: middle;
+`;
