+++ components/svg/charts/index.js
@@ -1,0 +1,2 @@
+export { default as ChartBase } from "./ChartBase";
+export { default as ChartFigure } from "./ChartFigure";
